requirements.txt
absl-py==2.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorlover==0.3.0
comm==0.1.4
contourpy==1.2.0
cryptography==43.0.0
cufflinks==0.17.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.18.1
flatbuffers==24.3.25
fonttools==4.47.0
fqdn==1.5.1
frozendict==2.3.8
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
imbalanced-learn==0.12.3
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lxml==4.9.3
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mplfinance==0.12.10b0
multitasking==0.0.11
namex==0.0.8
nbclient==0.8.0
nbconvert==7.9.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
NorenRestApiPy @ file:///Users/shambhunathsahu/Desktop/Shoonya%20API%20Bot/ShoonyaApi-py-master/dist/NorenRestApiPy-0.0.22-py2.py3-none-any.whl#sha256=ec7d35dd671ee919ad16bb158fdbe51322444a0fad2acc60a24e45dd7c1bf025
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.1
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.6
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandas_ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.16.1
pyotp==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.5.0
scikit-multilearn==0.2.0
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.45
stack-data==0.6.3
starlette==0.37.2
TA-Lib==0.4.28
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.11.1
typer==0.12.3
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.6
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wrapt==1.16.0
xgboost==2.1.0
XlsxWriter==3.2.0
yfinance==0.2.31
